requirements/dev.txt
-r base.txt
django-extensions==1.7.2